connector/requirements.txt
# Fivetran Custom Connector Dependencies

# Fivetran SDK and API
fivetran-connector-sdk>=2.2.1
fivetran-api-playground>=1.1.1

# Core HTTP and networking
requests>=2.32.5
urllib3>=2.5.0
httpx>=0.28.1

# Date and time handling
python-dateutil>=2.9.0
pytz>=2025.2

# JSON handling and data validation
jsonschema>=4.25.1
pydantic>=2.12.0
orjson>=3.11.3

# Logging and monitoring
structlog>=25.4.0
loguru>=0.7.3

# Retry and backoff capabilities
backoff>=2.2.1
tenacity>=9.1.2

# Environment and configuration management
python-decouple>=3.8
python-dotenv>=1.1.1

# Data processing utilities
pandas>=2.3.3
numpy>=2.3.3
openpyxl>=3.1.5

# Database connectivity (common for connectors)
sqlalchemy>=2.0.43
psycopg2-binary>=2.9.10
pymongo>=4.15.3

# Cloud storage support
boto3>=1.40.47
azure-storage-blob>=12.26.0
google-cloud-storage>=3.4.0

# Additional utilities
click>=8.3.0
tqdm>=4.67.1
rich>=14.1.0

# Security and encryption
cryptography>=46.0.2

# FastAPI and web server
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pyotp>=2.9.0

# Testing dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
requests-mock>=1.11.0
responses>=0.23.0

# Development and code quality
black>=23.9.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.5.0
pre-commit>=3.4.0

# Type hints and stubs
types-requests>=2.31.0
types-python-dateutil>=2.8.0